+++ lib/intervalBuilder.js
@@ -75,15 +75,15 @@ IntervalBuilder.prototype.changeObjectState = function (objectId, message) {
 };
 
 IntervalBuilder.prototype.isUp = function (objectId) {
-    return this.states[objectId] == UP;
+    return this.states[objectId] === UP;
 };
 
 IntervalBuilder.prototype.isDown = function (objectId) {
-    return this.states[objectId] == DOWN;
+    return this.states[objectId] === DOWN;
 };
 
 IntervalBuilder.prototype.isPaused = function (objectId) {
-    return this.states[objectId] == PAUSED;
+    return this.states[objectId] === PAUSED;
 };
 
 IntervalBuilder.prototype.build = function (begin, end, callback) {
@@ -93,17 +93,22 @@ IntervalBuilder.prototype.build = function (begin, end, callback) {
         initialState: function (next) {
             self.determineInitialState(begin, next);
         },
-        interval: ['initialState', function (next) {
+        interval: ['initialState', function (options, next) {
             self.buildIntervalsForPeriod(begin, end, next);
         }],
-        check: ['initialState', 'interval', function (next) {
-            self.getCurrentCheck(next)
+        check: ['initialState', 'interval', function (options, next) {
+            self.getCurrentCheck(options, next);
         }],
-        duration: ['initialState', 'interval', 'check', function (next, results) {
-            self.calculateDuration(results.check, begin, end, next);
+        duration: ['initialState', 'interval', 'check', function (options, next) {
+            if (typeof options.check !== 'undefined' && options.check !== null) {
+                self.calculateDuration(options.check, begin, end, next);
+            }
         }],
-        downtime: ['initialState', 'interval', 'check', 'duration', function (next, results) {
-            self.calculateDowntime(results.check, begin, end, next);
+        downtime: ['initialState', 'interval', 'check', 'duration', function (options, next) {
+            if (typeof options.check !== 'undefined' && options.check !== null) {
+                self.calculateDowntime(options.check, begin, end, next);
+            }
+            ;
         }]
     }, function (err, results) {
         if (err) {
@@ -117,23 +122,45 @@ IntervalBuilder.prototype.build = function (begin, end, callback) {
     });
 };
 
-IntervalBuilder.prototype.getCurrentCheck = function (callback) {
+
+IntervalBuilder.prototype.getCurrentCheck = function (options, callback) {
     Check.findOne({_id: this.objectIds[0]}, callback);
 };
 
 IntervalBuilder.prototype.determineInitialState = function (timestamp, callback) {
     var self = this;
+    // async.forEach(this.objectIds, function (objectId, next) {
+    //     CheckEvent.find()
+    //         .where('check').equals(objectId)
+    //         .where('timestamp').lte(timestamp)
+    //         .sort({timestamp: -1})
+    //         .findOne(function (err, checkEvent) {
+    //             if (err) return next(err);
+    //             if (!checkEvent) {
+    //                 // No ping ever - start the period as paused
+    //                 self.changeObjectState(objectId, 'paused', timestamp);
+    //                 return next();
+    //             }
+    //             self.changeObjectState(objectId, checkEvent.message);
+    //             next();
+    //         });
+    // }, function (err) {
+    //     if (err) return callback(err);
+    //     self.updateCurrentState(timestamp);
+    //     return callback(null);
+    // });
 
     // find the first event before the timestamp for a given check
-    var getInitialEvent = function (checkId, timestamp, callback) {
+    var getInitialEvent = function (objectId, timestamp, callback) {
         CheckEvent.find()
-            .where('check').equals(checkId)
+            .where('check').equals(objectId)
             .where('timestamp').lte(timestamp)
             .sort({timestamp: -1})
             .findOne(function (err, checkEvent) {
                 if (err) return callback(err);
                 if (!checkEvent) {
                     // No ping ever - start the period as paused
+                    // self.changeObjectState(objectId, 'paused', timestamp);
                     return callback(null, 'paused');
                 }
                 callback(null, checkEvent.message);
@@ -149,75 +176,9 @@ IntervalBuilder.prototype.determineInitialState = function (timestamp, callback)
         });
     }, function (err) {
         if (err) return callback(err);
-        self.setGlobalStateAtTime(self.getGlobalState(), timestamp);
+        self.updateCurrentState(timestamp);
         return callback(null, self.currentState);
     });
-};
-
-/**
- * Get the current state. If there is only one target, it's the target state.
- * Otherwise, the global state is:
- *  - down when at least one target is down
- *  - paused when all targets are paused
- *  - up otherwise
- */
-IntervalBuilder.prototype.getGlobalState = function () {
-    if (!this.isMultiTarget()) {
-        return this.states[this.objectIds[0]];
-    }
-    var paused = 0;
-    for (var objectId in this.states) {
-        if (this.isDown(objectId)) return DOWN; // at least one of the targets is down
-        if (this.isPaused(objectId)) paused++;
-    }
-    // multi-target states are paused only if all checks are paused
-    return paused === this.nbObjects ? PAUSED : UP;
-};
-
-/**
- * Set the global state for a given time, and update intervals if it's
- * different than the previous state.
- *
- * Return true if the global state was updated, false otherwise.
- */
-IntervalBuilder.prototype.setGlobalStateAtTime = function (globalState, timestamp) {
-    if (globalState === this.currentState) {
-        return false;
-    }
-    timestamp = this.getTimestamp(timestamp);
-    // complete the previous interval
-    this.completeCurrentInterval(timestamp);
-    // start a new interval
-    this.currentInterval = (globalState == UP) ? [] : [timestamp];
-    this.currentState = globalState;
-    return true;
-};
-
-IntervalBuilder.prototype.completeCurrentInterval = function (timestamp) {
-    if (this.currentInterval.length != 1) {
-        // no current interval - ignore
-        return false;
-    }
-    this.intervals.push(this.currentInterval.concat([timestamp, this.currentState]));
-    return true;
-};
-
-IntervalBuilder.prototype.buildIntervalsForPeriod = function (begin, end, callback) {
-    var self = this;
-    CheckEvent.find()
-        .where('check').in(this.objectIds)
-        .where('timestamp').gt(begin).lte(end)
-        .sort({timestamp: 1})
-        .find(function (err, checkEvents) {
-            if (err) return callback(err);
-            checkEvents.forEach(function (checkEvent) {
-                if (self.changeObjectState(checkEvent.check, checkEvent.message)) {
-                    self.setGlobalStateAtTime(self.getGlobalState(), checkEvent.timestamp);
-                }
-            });
-            self.completeCurrentInterval(end);
-            callback(null);
-        });
 };
 
 IntervalBuilder.prototype.calculateDuration = function (check, begin, end, callback) {
@@ -231,14 +192,14 @@ IntervalBuilder.prototype.calculateDuration = function (check, begin, end, callb
     }
 
     this.intervals.forEach(function (interval) {
-        if (interval[2] != PAUSED || interval[1] < durationBegin || interval[0] > durationEnd) {
+        if (interval[2] !== PAUSED || interval[1] < durationBegin || interval[0] > durationEnd) {
             return;
         }
 
-        var effectiveBegin = Math.max(durationBegin, interval[0]),
-            effectiveEnd = Math.min(durationEnd, interval[1]);
+        interval[0] = Math.max(durationBegin, interval[0]);
+        interval[1] = Math.min(durationEnd, interval[1]);
 
-        duration -= effectiveEnd - effectiveBegin;
+        duration -= interval[1] - interval[0];
     });
 
     return callback(null, duration);
@@ -250,21 +211,21 @@ IntervalBuilder.prototype.calculateDowntime = function (check, begin, end, callb
 
     if (this.isMultiTarget()) {
         // it's a tag - no other way
-        return callback(null, this.currentState == DOWN ? durationEnd - durationBegin : 0);
+        return callback(null, this.currentState === DOWN ? durationEnd - durationBegin : 0);
     }
 
     var downtime = 0;
     var currentIntervalEnd = null;
     this.intervals.forEach(function (interval) {
         currentIntervalEnd = interval[1];
-        if (interval[2] != DOWN || interval[1] < durationBegin || interval[0] > durationEnd) {
+        if (interval[2] !== DOWN || interval[1] < durationBegin || interval[0] > durationEnd) {
             return true; // will act as a continue
         }
 
-        var effectiveBegin = Math.max(durationBegin, interval[0]),
-            effectiveEnd = Math.min(durationEnd, interval[1]);
+        interval[0] = Math.max(durationBegin, interval[0]);
+        interval[1] = Math.min(durationEnd, interval[1]);
 
-        downtime += effectiveEnd - effectiveBegin;
+        downtime += interval[1] - interval[0];
     });
 
     if (!check.isUp) {
@@ -278,8 +239,90 @@ IntervalBuilder.prototype.calculateDowntime = function (check, begin, end, callb
     return callback(null, downtime);
 };
 
+IntervalBuilder.prototype.buildIntervalsForPeriod = function (begin, end, callback) {
+    var self = this;
+    CheckEvent.find()
+        .where('check').in(this.objectIds)
+        .where('timestamp').gt(begin).lte(end)
+        .sort({timestamp: 1})
+        .find(function (err, checkEvents) {
+            if (err) return callback(err);
+            checkEvents.forEach(function (checkEvent) {
+                if (self.changeObjectState(checkEvent.check, checkEvent.message)) {
+                    self.updateCurrentState(checkEvent.timestamp);
+                }
+            });
+            if (typeof callback === "function") {
+                self.completeCurrentInterval(end);
+                callback(null);
+            }
+        });
+};
+
+/**
+ * Set the global state for a given time, and update intervals if it's
+ * different than the previous state.
+ *
+ * Return true if the global state was updated, false otherwise.
+ */
+IntervalBuilder.prototype.updateCurrentState = function (timestamp) {
+    timestamp = this.getTimestamp(timestamp);
+    var currentState = this.getGlobalState();
+    if (currentState !== this.currentState) {
+        this.completeCurrentInterval(timestamp);
+        this.currentInterval = currentState === UP ? [] : [timestamp];
+        this.currentState = currentState;
+        return true;
+    }
+    return false;
+};
+
+// IntervalBuilder.prototype.updateCurrentState = function(timestamp) {
+//   timestamp = this.getTimestamp(timestamp);
+//   var currentState = this.getGlobalState();
+//   if (currentState === this.currentState) {
+//     return false;
+//   }
+//   this.completeCurrentInterval(timestamp);
+//   this.currentInterval = currentState == 1 ? [] : [timestamp];
+//   this.currentState = currentState;
+//   return true;
+// };
+
+IntervalBuilder.prototype.getGlobalState = function () {
+    var ups = 0;
+    var paused = 0;
+    for (var objectId in this.states) {
+        if (this.isUp(objectId)) ups++;
+        if (this.isPaused(objectId)) paused++;
+    }
+    if (!this.isMultiTarget() && paused > 0) {
+        return PAUSED; // global state is paused because there is only one check
+    }
+    if (this.isMultiTarget() && paused === this.nbObjects) {
+        return PAUSED; // global state is paused as all checks are paused
+    }
+    if ((ups + paused) === this.nbObjects) {
+        return UP; // ignore paused in multiTarget
+    }
+    return DOWN; // at least one of the targets is paused
+};
+
 IntervalBuilder.prototype.getTimestamp = function (date) {
     return date.valueOf ? date.valueOf() : date;
+};
+
+IntervalBuilder.prototype.completeCurrentInterval = function (timestamp) {
+    if (this.currentInterval.length !== 1) {
+        // no current interval - ignore
+        return false;
+    }
+    if (this.currentState === DOWN) {
+        this.downtime += timestamp - this.currentInterval[0];
+    }
+
+    this.intervals.push(this.currentInterval.concat([timestamp, this.currentState]));
+    return true;
 };
 
 module.exports = IntervalBuilder;
